write dynamic vehicle update query

+++ src/config/db.ts
@@ -24,7 +24,7 @@ const initDB = async () => {
             vehicle_name VARCHAR(100) NOT NULL,
             type VARCHAR(50),
             registration_number VARCHAR(50) UNIQUE NOT NULL,
-            daily_rent_price DECIMAL(10, 2) NOT NULL,
+            daily_rent_price INT NOT NULL,
             availability_status VARCHAR(50) DEFAULT 'available'
         )
         `);
@@ -36,7 +36,7 @@ const initDB = async () => {
             vehicle_id INT REFERENCES vehicles(id) ON DELETE CASCADE,
             rent_start_date DATE NOT NULL,
             rent_end_date DATE NOT NULL CHECK (rent_end_date > rent_start_date),
-            total_price DECIMAL(10, 2) NOT NULL,
+            total_price INT NOT NULL,
             status VARCHAR(50)
         )
         `);
